+++ src/parser/ELR/model/re-lex.ts
@@ -5,7 +5,7 @@ import { Callback, GrammarRuleContext } from "./context";
 export type ReLexStack<State, T> = {
   readonly stateStack: State[];
   readonly buffer: ASTNode<T>[];
-  readonly lexer: ILexer<any>; // TODO: use generic type
+  readonly lexer: ILexer<any, any>; // TODO: use generic type
   readonly index: number;
   readonly errors: ASTNode<T>[]; // TODO: is this needed?
   readonly rollbackStackLength: number;