+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   - Feat: add `ILexer.takeUntil`. [#16](https://github.com/DiscreteTom/retsac/issues/16)
   - Feat: add `Lexer.utils.regexLiteral`. [#15](https://github.com/DiscreteTom/retsac/issues/15)
   - Feat: add `ILexer.lex.peek`.
+  - Feat: typed lexer. [#17](https://github.com/DiscreteTom/retsac/issues/17)
 - Parser
   - Feat: add `name` for `ASTNode` and `ASTObj`.
     - **_Breaking Change_**: `ASTNode.toTreeString` will also print `ASTNode.name`.